requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pkg-resources==0.0.0
PyYAML==5.3.1
requests==2.23.0
tqdm==4.43.0
urllib3==1.25.8
